scripts/lpp_collector/pyproject.toml
[tool.poetry]
name = "lpp-collector"
version = "0.1.0"
description = ""
authors = ["f0reachARR <f0reach@f0reach.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^6.2.5"
whiptail-dialogs = "^0.4.0"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[tool.poetry.plugins.pytest11]
lpp_collector = "lpp_collector"

[tool.poetry.scripts]
consent = "lpp_collector.consent:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
